Process results within af function

This function is used by the af function to process the results when iterating over different boundary values

process.fn(fence.mod, fence.rank)

Arguments

fence.mod
set of fence models

fence.rank
set of fence model ranks